(Refact) seperate map from page
